move date resolution to AbstractFile
